+++ app/endpoints/profile.py
@@ -1,8 +1,10 @@
-from fastapi import APIRouter, Form, Request
+from fastapi import APIRouter, Request, Depends
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
+from sqlalchemy.ext.asyncio import AsyncSession
 
-from db import profile_crud, grade_types_crud, work_types_crud
+from db import get_async_session, profile_crud, grade_types_crud, work_types_crud
+from db.schemas import ProfileCreate
 from .validators import validate_salary
 
 
@@ -14,12 +16,13 @@ templates = Jinja2Templates(directory='app/templates')
 async def get_profile_form(
     request: Request,
     user_id: int,
+    session: AsyncSession = Depends(get_async_session),
 ):
     """
     returns profile template.
     """
     profile = await profile_crud.get_by_attribute(
-        'user_id', user_id, is_deleted=False
+        'user_id', user_id, session, is_deleted=False
     )
     if profile:
         return templates.TemplateResponse(
@@ -31,37 +34,25 @@ async def get_profile_form(
 @router.post('/{user_id}')
 async def create_or_update_profile(
     user_id: int,
-    professional_role: str = Form(..., min_length=1, max_length=32),
-    grade: str = Form(...),
-    work_type: str = Form(...),
-    region: str = Form(..., min_length=1, max_length=32),
-    salary_from: int = Form(..., gt=0),
-    salary_to: int = Form(..., gt=0),
-    ready_for_relocation: bool = Form(False),
+    form: ProfileCreate = Depends(ProfileCreate.as_form),
+    session: AsyncSession = Depends(get_async_session),
 ) -> dict[str: str | int | bool]:
     """
     creates new profile or updates existing.
     """
-    validate_salary(salary_from, salary_to)
-    grade = await grade_types_crud.get_by_attribute('type', grade)
-    work_type = await work_types_crud.get_by_attribute('type', work_type)
-    data = {
-        "professional_role": professional_role,
-        "grade_type_id": grade.id,
-        "work_type_id": work_type.id,
-        "region": region,
-        "salary_from": salary_from,
-        "salary_to": salary_to,
-        "ready_for_relocation": ready_for_relocation,
-        "user_id": user_id,
-    }
+    data = form.dict()
+    validate_salary(data['salary_from'], data['salary_to'])
+    grade = await grade_types_crud.get_by_attribute('type', data.pop('grade'), session)
+    work_type = await work_types_crud.get_by_attribute('type', data.pop('work_type'), session)
+    data['grade_type_id'] = grade.id
+    data['work_type_id'] = work_type.id
+    data['user_id'] = user_id
 
     profile = await profile_crud.get_by_attribute(
-        'user_id', user_id, is_deleted=False
+        'user_id', user_id, session, is_deleted=False
     )
-
     if profile:
-        await profile_crud.update(profile, data)
+        await profile_crud.update(profile, data, session)
     else:
-        await profile_crud.create(data)
+        await profile_crud.create(data, session)
     return data
