+++ db/crud/base.py
@@ -1,15 +1,15 @@
 from typing import Optional
 
 from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
 
-from db.engine import AsyncSessionLocal
-from db.models import Profile, GradeTypes, WorkTypes, User, BaseModel
+from db.models.profile import Profile, GradeTypes, WorkTypes
+from db.models.user import User
+from db.models.base import BaseModel
 
 
 class CRUDBase:
-    """
-    Basic CRUD operaions to work with db.
-    """
+    """Базовые CRUD операции для работы с БД."""
 
     def __init__(self, model):
         self.model = model
@@ -17,106 +17,87 @@ class CRUDBase:
     async def get_by_id(
             self,
             obj_id: int,
+            session: AsyncSession,
     ) -> BaseModel:
-        """
-        Returns db_object by it's id value.
-        """
-        async with AsyncSessionLocal() as session:
-            db_obj = await session.execute(
-                select(self.model).where(
-                    self.model.id == obj_id
-                )
+        """Возвращает объект по его id."""
+        db_obj = await session.execute(
+            select(self.model).where(
+                self.model.id == obj_id
             )
-            obj = db_obj.scalars().first()
-        return obj
+        )
+        return db_obj.scalars().first()
 
     async def get_by_attribute(
             self,
             attr_name: str,
             attr_value: str | int | bool,
+            session: AsyncSession,
             is_deleted: Optional[bool] = None,
     ) -> BaseModel:
-        """
-        Returns db_object by any attr value.
-        """
         attr = getattr(self.model, attr_name)
 
         if is_deleted is not None and (
             self.model == Profile or self.model == User
         ):
-            async with AsyncSessionLocal() as session:
-                db_obj = await session.execute(
-                    select(self.model).where(
-                        attr == attr_value, self.model.is_deleted == is_deleted
-                    )
-                )
-                obj = db_obj.scalars().first()
-            return obj
-
-        async with AsyncSessionLocal() as session:
             db_obj = await session.execute(
-                select(self.model).where(attr == attr_value)
+                select(self.model).where(
+                    attr == attr_value, self.model.is_deleted == is_deleted
+                )
             )
-            obj = db_obj.scalars().first()
-        return obj
+            return db_obj.scalars().first()
+
+        db_obj = await session.execute(
+            select(self.model).where(attr == attr_value)
+        )
+        return db_obj.scalars().first()
 
     async def get_multi(
             self,
+            session: AsyncSession
     ) -> list[BaseModel]:
-        """
-        Returns list of db_objects
-        """
-        async with AsyncSessionLocal() as session:
-            db_objs = await session.execute(select(self.model))
-            objs = db_objs.scalars().all()
-        return objs
+        """Возвращает список всех объектов модели."""
+        db_objs = await session.execute(select(self.model))
+        return db_objs.scalars().all()
 
     async def create(
             self,
             obj_in_data: dict[str: str | int | bool],
+            session: AsyncSession,
     ) -> None:
-        """
-        Create new db_object and returns it.
-        """
         db_obj = self.model(**obj_in_data)
-        async with AsyncSessionLocal() as session:
-            session.add(db_obj)
-            await session.commit()
-            await session.refresh(db_obj)
-            return db_obj
+        session.add(db_obj)
+        await session.commit()
+        # await session.refresh(db_obj)
+        # return db_obj
 
     async def update(
             self,
             db_obj: BaseModel,
             obj_in_data: dict[str: str | int | bool],
+            session: AsyncSession,
     ) -> None:
-        """
-        Updates db_object and returns refreshed one.
-        """
+        """Вносит изменения в объект базы данных."""
+
         for field, value in obj_in_data.items():
             setattr(db_obj, field, value)
-        async with AsyncSessionLocal() as session:
-            session.add(db_obj)
-            await session.commit()
-            await session.refresh(db_obj)
-            return db_obj
+        session.add(db_obj)
+        await session.commit()
+        # await session.refresh(db_obj)
+        # return db_obj
 
     async def delete(
             self,
             db_obj: BaseModel,
+            session: AsyncSession,
     ) -> None:
-        """
-        Sets is_deleted status to deleted db_objects.
-        """
+        """Отмечает объект из базы данных как удаленный."""
         setattr(db_obj, "is_deleted", True)
-        async with AsyncSessionLocal() as session:
-            session.add(db_obj)
-            if self.model == User:
-                setattr(db_obj.profile, "is_deleted", True)
-                session.add(db_obj.profile)
-            await session.commit()
-            await session.refresh(db_obj)
-            return db_obj
+        session.add(db_obj)
+        if self.model == User:
+            setattr(db_obj.profile, "is_deleted", True)
+            session.add(db_obj.profile)
+        await session.commit()
+        # return db_obj
 
 
 user_crud = CRUDBase(User)
