+++ db/crud/profile.py
@@ -2,9 +2,9 @@ from typing import Optional
 
 from sqlalchemy import select
 from sqlalchemy.orm import joinedload
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from db.crud.base import CRUDBase
-from db.engine import AsyncSessionLocal
 from db.models import Profile
 
 
@@ -16,6 +16,7 @@ class CRUDProfile(CRUDBase):
             self,
             attr_name: str,
             attr_value: str | int | bool,
+            session: AsyncSession,
             is_deleted: Optional[bool] = None,
     ) -> Profile:
         """
@@ -23,16 +24,15 @@ class CRUDProfile(CRUDBase):
         """
         attr = getattr(self.model, attr_name)
         if is_deleted is not None and self.model == Profile:
-            async with AsyncSessionLocal() as session:
-                db_obj = await session.execute(
-                    select(self.model).options(
-                        joinedload(Profile.work_type),
-                        joinedload(Profile.grade_type)).where(
-                            attr == attr_value,
-                            self.model.is_deleted == is_deleted
-                        )
+            db_obj = await session.execute(
+                select(self.model).options(
+                    joinedload(Profile.work_type),
+                    joinedload(Profile.grade_type)).where(
+                        attr == attr_value,
+                        self.model.is_deleted == is_deleted
                     )
-                return db_obj.scalars().first()
+                )
+            return db_obj.scalars().first()
 
 
 profile_crud = CRUDProfile(Profile)
