+++ db/engine.py
@@ -19,7 +19,8 @@ def create_async_engine(url: URL | str) -> AsyncEngine:
     :param url:
     :return:
     """
-    return _create_async_engine(url=url, echo=True, pool_pre_ping=True)
+    return _create_async_engine(url=url, pool_pre_ping=True)
+    # return _create_async_engine(url=url, echo=True, pool_pre_ping=True)
 
 
 async def get_session_maker(engine: AsyncEngine) -> async_sessionmaker:
