+++ requirements.txt
@@ -16,6 +16,7 @@ colorama==0.4.6
 fastapi==0.78.0
 frozenlist==1.3.3
 greenlet==2.0.2
+gunicorn==20.1.0
 h11==0.14.0
 httptools==0.5.0
 idna==3.4
